requirements.txt
beautifulsoup4==4.12.0
Flask==2.2.3
unidecode==1.3.6
requests==2.28.2
nltk==3.8.1
python-dotenv==1.0.0